Add S3 syncing script
This is a much simpler replacement for flatpak-export-to-s3.py from
eos-build. Rather than write our own uploader in boto, aws s3 is used,
which is multithreaded and has logic for when to sync files. With that,
we can simply sync the entire s3/screenshots and s3/thumbnails
directories all the time and not have to worry about filtering.

The eos-extra.xml.gz file still has to be created. It's done locally in
the s3/app-info directory with a bit logic to only overwrite it when
necessary.

The intention is to have this run by a jenkins job after a git push to
this repo.

https://phabricator.endlessm.com/T18125