Remove all existing screenshots
Most of these were generated incorrectly, and we are about to
replace them with the proper version, so let's start with a
clean slate.  Since the generator script does not automatically
remove those assets that are no longer needed, we want to remove
the old ones so as not to bloat the image builds.

https://phabricator.endlessm.com/T19914