eos-extra.xml: Merge in popular apps list from gnome-software eos plugin
Previously, the list of featured apps to present in gnome-software in
Endless OS was hardcoded in the `eos` plugin in our fork of
gnome-software. That’s not very maintainable, and means that the list of
popular apps can only be updated once per OS release.

Instead, provide the list using the existing eos-extra.xml file, which
is set in gnome-software’s `external-appstream-urls` list and which
gnome-software will update automatically when connected to the internet.

Some of these apps were in eos-extra.xml already; in those cases, add
the kudos to the existing component element.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T26507