eos-extra.xml: Specify popular apps as flatpak bundles
GNOME Software has a mechanism to hand over GsApps created by
appstream to other plugins. However, in order to adopt an app[1],
bundle type should be mentioned in the appstream format. This is
how the flatpak plugin can take over the GsApp created by Appstream
plugin (apps mentioned in eos-extra.xml are parsed by appstream plugin
first and then handed over to the flatpak plugin).

The "bundle" XML node, can remain empty, which gets be filled in
by the flatpak plugin during it's refine.

Also, as of gnome-software 3.32, flatpak plugin strips ".desktop"
from <id> node of the appstream file; whatever remains is the
flatpak ref for the app. Unfortunately, we cannot replicate that logic
in the appstream plugin (remember that, apps in eos-extra.xml are
created by appstream plugin first) so we just mention the flatpak
ref as <id> in eos-extra.xml without the ".desktop" suffix.

This workaround enables the flatpak plugin to:
1. Adopt the eos-extra.xml originated GsApp
2. Refine the GsApp adopted in 1., using the flatpak repo's appstream
   data.
3. Respect additional metadata (like <kudos> mentioned) in eos-extra.xml

[1] https://developer.gnome.org/gnome-software/stable/plugins.html

https://phabricator.endlessm.com/T26507