eos-extra: Specify ref for each application
Earlier the refs were not specified in eos-extra because of their
arch-dependent nature. As per [1], gnome-software can resolve
the default arch for the current platform.

[1] https://gitlab.gnome.org/GNOME/gnome-software/merge_requests/297

After consulting with upstream maintainers, they specified that the
 "origin" property for external-appstream is not necessary although
the appstream standard mentions it as a mandatory property.
As eos-extra comprises of apps belonging to different remote, the
"origin" for that specific app can be computed by gnome-software
itself. Hence, we drop the "origin" property from this external
appstream file.

https://phabricator.endlessm.com/T26507