app-info: Regenerate eos-apps XML from script
This is the result of running `generate_endless_apps_popular_xml.sh` and
copying the output into `eos-extra.xml`.

It adds the bundle IDs to all the eos-apps entries, and one new app
(`com.endlessm.encyclopedia.ru`) which has been added to the repository
since the XML was last updated.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

https://phabricator.endlessm.com/T30217