eos-extra: Refresh popular apps
The new list is based on a combination of:

- Apps that have "Get ____" launchers from eos-application-tools in
  eos3.8
- Metrics on popular apps
- Manual review by me and Rob McQueen
- Previous work on prioritizing apps to preinstall[0], which was itself
  based on metrics and manual review

Some particular notes on this new list:

- We no longer feature any built-in apps. Previously, we did this
  because not all apps were shown on the desktop. Now, all apps are
  shown on the desktop, so the app center is more about discovering new
  apps to install.
- We feature popular third-party apps, measured in part by metrics, and
  particularly those which we can't pre-install. It is likely that we
  will remove the "Get ____" launchers we previously used to work around
  not being able to pre-install certain popular apps.
- We no longer feature all com.endlessm.* apps. We now feature a
  somewhat-curated selection of the more popular apps. We do not feature
  apps whose nightly ingesters are broken.
- We no longer feature all com.endlessnetwork.* apps. We feature all the
  games, plus the most popular of the reference apps.
- We feature Kolibri, a key part of our offline content strategy.

The new list is rather shorter than the old list, but it is still more
than 35 apps, which is the maximum number shown in the App Center on any
given day.

As part of reviewing this list, I implemented a small tool to turn a
list of Flatpak app IDs into the required AppStream XML.  (It is a nice
consequence of not featuring any built-in apps that we can do this,
though that was not the reason!)

One thing I do not understand is: in the flathub and eos-apps appstream,
many apps have (eg):

  <component type="desktop">
    <id>com.endlessm.encyclopedia.en.desktop</id>
                                    ^^^^^^^^

If I include the '.desktop' in this eos-extra.xml file, then the app is
NOT marked as popular. If I remove the '.desktop', as is done by this
script, then it is.  I find this deeply unsatisfactory, particularly
since I went to some length to read the primary appstream file for each
remote and determine the appstream ID by looking up the Flatpak bundle
rather than assuming it does or does not have .desktop at the end.

I've discarded the custom categories previously added here. I have to
admit that I have not reviewed them all to make sure they are present in
the upstream appdata.

I've also discarded any custom background images.

https://phabricator.endlessm.com/T30820
https://phabricator.endlessm.com/T30379
https://phabricator.endlessm.com/T30764

[0]: https://docs.google.com/spreadsheets/d/1VTuLuF2o3FOYAq69MkdNvzy5D9G6n4DeuUVv7wUtlE4/