app-info: Feature apps in more places
Add more things to the generated metainfo which means that the apps
should be featured in more places. Previously, they were only shown in
the “Editor’s Choice” section of the overview page. Now they should
additionally be shown in “Editor’s Choice” on relevant category pages,
and in the featured carousel on both the overview and category pages.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

https://phabricator.endlessm.com/T33043