sync-s3.sh: Invalidate CloudFront CDN
When files are updated in the S3 bucket, clients may not get them if
CloudFront is still caching them. Take the brute force approach and
invalidate the entire distribution to ensure any updates reach users. It
would be better to only invalidate the updated paths, but that would
involve a much bigger rewrite.

https://phabricator.endlessm.com/T33208