pwa-metainfo-generator.py: Switch from CSV input to YAML
CSV is reasonable as a data format, but it's not great as a
configuration format. As the generator gains more options to adjust its
output, CSV starts to become a hindrance. YAML provides several features
that weren't really possible with CSV:

- Truly optional fields
- Data types beyond strings
- Structured data like sequences and maps
- Multi-line strings
- Comments

This has been tested with a conversion of `gnome-pwa-list-foss.csv` and
produced identical output. Two of the field names were changed.
`is_adaptive` is now just `adaptive` since it can clearly be expressed
as a boolean. `custom_summary` is just `summary` since I think it's
clear that you're providing a custom summary in that field.

https://phabricator.endlessm.com/T34477