pwa-metainfo-generator.py: Reorder metainfo fields
This is obviously highly subjective, but I believe that the app ID and
name should be at the beginning of the entry with the summary following
shortly. The order doesn't matter to AppStream clients, but I think this
makes the generated XML easier to review. Since the order of the YAML
fields also doesn't matter, reorder the help text so it corresponds with
the order of the generated fields.

https://phabricator.endlessm.com/T34477