pwa-metainfo-generator.py: Add optional screenshots support
If the manifest does not provide screenshots or they are not ideal,
allow providing screenshots in the configuration. The configuration
fields roughly match the manifest fields to ease the implementation.

Later support for thumbnails can be added. It would also be nice if the
configured screenshots could augment the manifest's screenshots instead
of simply replacing them.

https://phabricator.endlessm.com/T34477