app-info: Change catalog origin to eos-extra
The specification suggests that `origin` should match the basename of
the filename[1]. Until now this has almost certainly not mattered as all
of the entries used `merge="append"`, which means they were just adding
additional metadata to an existing entry. Furthermore, the spec suggests
this is used for icon caching, and the existing entries don't specify
any icons.

Now we're going to be adding web application entries where we are both
the primary metadata provider and are specifying icons. Set `origin` as
the spec suggests instead of using `flathub`, which would be wrong.

1. https://www.freedesktop.org/software/appstream/docs/chap-CatalogData.html#spec-asxml-general

https://phabricator.endlessm.com/T34477