pwa-metainfo-generator: Change to generating XML skeletons
As we strive to publish rich PWA appstream, there will always be cases
where it can't be regenerated from configuration and heuristics. Instead
of trying to encode every decision into YAML, scrap the whole concept
and have the generator dump out all the information it can find in a
metainfo XML skeleton.

It will be up to the editor to provide the best appstream data, but once
it's there it will be maintained in the metainfo XML until someone wants
to change it again. The generator's job will be to provide the upstream
data and guide you to produce the best possible metainfo. It can be run
again later (possibly in `--stdout` mode) to see if the website has
provided any new or updated metadata to be included in the metainfo.

https://phabricator.endlessm.com/T34536