app-info: Convert PWA YAML configuration to URL file
pwa-metainfo-generator.py no longer supports YAML configuration and the
desired state is already recorded in the metainfo XML files. Instead,
keep a simple URL file for the PWAs we've generated. This will serve
primarily as a record of our desired PWAs, but it will also allow
regenerating the metainfo periodically to see if there's any new data
upstream to use.

https://phabricator.endlessm.com/T34536