app-info: Add metainfo translation infrastructure
Instead of inserting the translated values directly as `xml:lang`
elements in the metainfo, setup translation infrastructure to do it.
This uses the metainfo ITS handling from gettext and itstool to do the
substitutions.

After this, the committed metainfo files are templates and
`eos-extra.xml` has to be generated after the substituting the
translations. A simple `Makefile` setup has been added to support that.
For now `eos-extra.xml` is still committed, but we may want to switch to
it being always being generated at build time.

https://phabricator.endlessm.com/T34536