pwa-metainfo-generator: Suggest translations in po files
Instead of adding the translated elements directly in the metainfo file,
suggest that they get added to the relevant po file. Otherwise, we'd
have to mark the element as untranslatable so that someone adding a
gettext translation didn't duplicate the directly added translation. In
other words, gettext is all or nothing here.

https://phabricator.endlessm.com/T34536